% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPSM_main.R
\name{psm_integrate}
\alias{psm_integrate}
\title{scPSM integration}
\usage{
psm_integrate(
  batches,
  markers,
  hvg,
  k.self = 10,
  k.mnn = 10,
  correct.all = TRUE,
  merge.order = 1:4
)
}
\arguments{
\item{batches}{One or more log-expression matrices where genes correspond to rows and cells correspond to columns.
Each matrix should contain the same number of rows, corresponding to the same genes in the same order.
Each matrix represents a batch.}

\item{markers}{A vector specifying which features used as marker genes to compute propensity scores.}

\item{hvg}{A vector specifying which features used as HVGs for identifying MNN group.}

\item{k.self}{An integer scalar specifying the number of nearest neighbors in searching KNNs.}

\item{k.mnn}{An integer scalar specifying the number of nearest neighbors in matching MNN pairs.}

\item{correct.all}{A logical scalar specifying whether correction should be applied to all genes, even if only a subset is used for the MNN group identification.}

\item{merge.order}{An integer vector containing the linear merge order of batches.}
}
\description{
Correct for batch effects, impute dropouts, and denoise data
using the propensity score matching method for single-cell RNA-sequencing data.
}
\examples{
#Please see the tutorial for the preprocessing and then run the following code
#psm.data <- psm_integrate(batches = batches, markers = markers, hvg = hvg, merge.order = 1:4)
}
